Fixing test waiter to use requestAnimationFrame when in test mode (#280)
In certain cases the test waiter never completes in tests when we're using whenRouteIdle. The hypothesis is that during tests, the main thread is pinned and never gets into an idle state, thus resulting in the requestIdleCallback to never fire.

This PR changes the semantics, during test only, of what schedule function is used. In test we will always fall back to `requestAnimationFrame` rather than `requestIdleCallback`.